\name{netcmc-package}
\alias{netcmc-package}
\alias{netcmc}
\docType{package}
\title{
  An R Package for Bayesian Social Network Modelling
}
\description{
 Implements a class of univariate and multivariate spatio-network generalised linear mixed 
models, with inference in a Bayesian setting using Markov chain 
Monte Carlo (MCMC) simulation. The response variable can be binomial, Gaussian, and Poisson.
}

\details{
\tabular{ll}{
Package: \tab netcmc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2022-01-24\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
George Gerogiannis <g.gerogiannis.1@research.gla.ac.uk>
}
\examples{
  ## See the examples in the function specific help files.
}
